Update 03-large-files.md
This is Suresh, Product Manager of File System squad. File System Team TL Doug & CSE Team Jordan & Myself contributed to the changes in this doc.

Major trigger for the change is the release of a feature (which will happen in this quarter) related to highly populated directories and this is my effort to ensure that the content is up to date.